fix tagging by detecting commit matching PKGBUILD file location
The issue is that the repos directory can change per archrelease commit
hence we cannot simply use a glob to determine the location but need to
get the location out of the commit object.
We unfortunately also cannot use the prefered diff-tree non porcelain
variant as that will not work for the root commit.