handle commits removing plus adding a PKGBUILD at the same time
archrelease may remove some PKGBUILD from a testing repo while releasing
one for another repo in the same transaction commit. Handle this by
trying to fetch any of the files touching a PKGBUILD.
Additionally only treat non fetchable PKGBUILD as error when we expect
that one must be available like during an "archrelease: copy" in
contrary to looping through all commits of the tree that may contain a
version that did not have any PKGBUILD inside the tree.